Auto-save notes when dismissing checkmark and number dialogs
Previously, when a user entered notes for a habit entry, then dismissed the
dialog (either intentionally or, as often the case for me, accidentally), any
changes to the notes would be lost.

This change seeks to avoid data loss by auto-saving notes when the dialog is
closed. This handles a several data loss scenarios:

- pressing the back button
- clicking outside the dialog box
- rotating the screen
- turning off the screen

(To be honest, I have done all of these at least once by accident, and some
dozens of times.)

As part of this change, the NumberDialog now trims whitespace from notes when
manually saving to be consistent with CheckmarkDialog.